requirements.txt
ffmpeg-python==0.2.0
numpy==1.23.4
scikit-learn==1.1.3
torch==1.13.0
transformers==4.24.0
whisper @ git+https://github.com/openai/whisper.git@9f70a352f9f8630ab3aa0d06af5cb9532bd8c21d
youtube-dl==2021.12.17
streamlit==1.14.0
pandas==1.5.1
altair==4.2.0
